requirements.txt
lightgbm==3.2.1
numpy==1.21.4
optuna==2.10.0
pandas==1.3.5
scikit-learn==1.0.2
scipy==1.7.3
statsmodels==0.13.1
pandas-gbq==0.17.5
google-cloud-storage==2.3.0
google-cloud-bigquery==3.1.0